Revert "Merge branch 'struct' into struct"
This reverts commit 66e90f5d229d105c6cc4c5b051bf84425632e970, reversing
changes made to b681b6c6c60c096fc0f207232c364dbadf7e7997.
