+++ src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
   {
     path: 'home',
-    loadComponent: () => import('./pages/home/home.page').then((m) => m.HomePage),
+    loadComponent: () => import('./home/home.page').then((m) => m.HomePage),
   },
   {
     path: '',
@@ -12,20 +12,20 @@ export const routes: Routes = [
   },
   {
     path: 'auth',
-    loadComponent: () => import('./pages/auth/auth.page').then( m => m.AuthPage)
+    loadComponent: () => import('./auth/auth.page').then( m => m.AuthPage)
   },
   {
     path: 'auth-sign-up',
-    loadComponent: () => import('./pages/auth-sign-up/auth-sign-up.page').then( m => m.AuthSignUpPage)
+    loadComponent: () => import('./auth-sign-up/auth-sign-up.page').then( m => m.AuthSignUpPage)
   },
   {
     path: 'auth-options',
-    loadComponent: () => import('./pages/auth-options/auth-options.page').then( m => m.AuthOptionsPage)
+    loadComponent: () => import('./auth-options/auth-options.page').then( m => m.AuthOptionsPage)
   },
 
   {
     path: 'profile',
-    loadComponent: () => import('./pages/profile/profile.page').then( m => m.ProfilePage)
+    loadComponent: () => import('./profile/profile.page').then( m => m.ProfilePage)
   }
 
 ];
