+++ src/app/home/home.page.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { IonHeader, IonToolbar, IonTitle, IonContent, IonMenu, IonMenuButton, IonButtons } from '@ionic/angular/standalone';
+
+
+@Component({
+  selector: 'app-home',
+  templateUrl: 'home.page.html',
+  styleUrls: ['home.page.scss'],
+  standalone: true,
+  imports: [IonButtons, IonHeader, IonToolbar, IonTitle, IonContent, IonMenu, IonMenuButton],
+})
+export class HomePage {
+  constructor() {}
+}
