--- src/app/pages/auth/auth.page.html
@@ -1,27 +1,0 @@
-<app-menu></app-menu>
-<div class="login-content"></div>
-<ion-content>
-  <form [formGroup]="loginForm" (ngSubmit)="onSubmit()" class="formControl">
-    <app-custom-input
-    [control]="loginForm.controls.email"
-    label="Email"
-    type="email"
-    autocomplete="email"
-  ></app-custom-input>
-
-  <app-custom-input
-    [control]="loginForm.controls.password"
-    label="Password"
-    type="password"
-    autocomplete="Password"
-  ></app-custom-input>
-
-  <ion-button
-    expand="block"
-    type="submit"
-    [disabled]="!loginForm.valid"
-    >
-    Click me
-  </ion-button>
-  </form>
-</ion-content>
