--- src/app/pages/home/home.page.html
@@ -1,8 +1,0 @@
-
-<ion-content [fullscreen]="true">
-  <app-menu></app-menu>
-  <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Blank</ion-title>
-    </ion-toolbar>
-  </ion-header>
