+++ src/app/profile/profile.page.scss
@@ -1,9 +1,3 @@
-.menu{
-  display: flex;
-  justify-content: center;
-  padding: 30px;
-
-}
 ion-menu-button {
   color: var(--ion-color-primary);
 }
